fixing lightbox still (mobile also) pass 6 - AG 0-2
Got swipe to close working (I think, will have to test on mobile once this pushes). Next steps are fixing the lightbox height since taking h-screen off, as the bottom was getting cut off, and seeing if we can mimic fancybox's horizontal swipe effect. I imagine that will be difficult, as we originally went with a css/html approach that kind of just worked (ArtistGallery3+). If we can mimic fancybox with js, then we are golden, but if we can't, we may have to revisit AG3+ (sigh). Lightboxes suck. Once this is done, if it's good, I'm going to export it to its own project, maybe a component library or something, so as to reuse it if I ever need to do this shit again. Web sucks, lol.