fixing lightbox still (mobile also) pass 14 - AG 0-5-11-4(pinchzoom) (pinch zoom pass 2)
Ongoing: incrementally develop out pinch zoom in depth: with fancybox as a reference, we are replicating meaningful behavior piece by piece in each installment of ArtistGallery. The order for pinch zoom specifically goes: ArtistGallery0-5-7 through 11, then backtracking to 10, and then on to 0-5-11-2 through 4 (4 being current as of this commit). Slowly but surely we are tightening up gesture behavior. Outstanding is preventing other gestures (3 finger, etc), getting buttons working, showing, and looking nicely, and lazy loading images with perhaps a waiting graphic of some kind, and then a good overall refactor of the component.