fixing lightbox still (mobile also) pass 14 - AG 0-5-11-8(pinchzoom - incremental deltas - using pendingRef in place of old refs) (pinch zoom pass 4)
This commit represents the latest of two competing versions of the pinchzoom feature, as computed via a baseline versus via incremental deltas. Neither are entirely refactored, but the latest files are 'deltas from a baseline' and 'incremental deltas - using pendingRef in place of old refs'. From a brief test, these two are functionally the same, but go about it differently. This commit will keep the history of the decision after which I will be going with one over the other, and working henceforth on the chosen one, which whill probably be the incremental deltas (the code is simpler). If at any point something comes up and this was the wrong decision, I'll have to retrofit the baseline version into whatever I'm doing at that point. I will probably go ahead and delete all the clutter versions now, including the baseline version, but I'll keep that in some back-pocket area as well as in this commit.